Stephen dev (#4)
* Test/Created Currency

Just a test to make sure this be pushing correct.

Created a script called Currency

* Basic VR Interactions

Grab objects, teleport, snap sockets, spawn socket

* More VR Setup

Refined snap socket prefabs, a socket now follows the horizontal position of a tower when held and sits .1 above the ground

* Raycast socket projection

The location of the socket is now based on a spherecast straight down from the block.  If it hits anything other than "Placeable Ground", it doesn't update.  For some reason, the spherecasts don't stop when the block is put down.

* Fixed snap locations & spawner

Removed all use of built in XR socket because it causes interaction issues.  Built own, specialized version.

* Currency System, Started Implementation of

Created a class that handles incrementing, decreasing, etc. of currency used by the player to play the game.

Created an object in-scene that displays current currency to user via TextMeshPro, and a script to handle the necessary updates of the display

Co-Authored-By: Sawyer Knowles <46873273+SawyerKnowles@users.noreply.github.com>
Co-Authored-By: Stephen Ogden <42942300+sjogden@users.noreply.github.com>
Co-Authored-By: FrankVanHove <46799346+FrankVanHove@users.noreply.github.com>

* Documentation

Added documentation to custom scripts.  Also added ability to set allowed layer name on spherecast from editor

* Integrated Currency Scripts to Tower Spawning

Towers now only spawn if the currency is above a configurable amount.  Also minor bugfixes to remove console warnings.

* Menu Scene Started

A menu scene is created with a VR rig in it.  There is a menu pane with 3 buttons on it that I can select with a ray pointer from the controllers, but they do not have actions attached yet.

* Added ability to assign actions to buttons

Script ButtonActions can now be added to panes to allow the association of Buttons to Actions.  This is configurable through the inspector via the instructions in the comments.

* Deleted problematic meta files

Co-authored-by: Maritza Gonzalez <59942843+mcgonzale@users.noreply.github.com>
Co-authored-by: Sawyer Knowles <46873273+SawyerKnowles@users.noreply.github.com>
Co-authored-by: FrankVanHove <46799346+FrankVanHove@users.noreply.github.com>