Implemented tower upgrade placing
Tower upgrades can now be placed (only) on top of existing towers.  Towers no longer have gravity at any point, due to a glitch that did not turn off gravity when instructed.
Created duplicates of the main scene and the tower prefab to avoid making changes to the main folder.